+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.2.2:
+
+* Add chef_integrate recipe
+ * But doesn't support starting chef-rundeck process
+
 ## v0.2.0:
 
 * Add tool_users recipe